Ensure admission webhook port is always a string
Before this commit, the port would be rendered in correctly after templating.
The value would look like :

```
	- name: CONTROLLER_ADMISSION_WEBHOOK_LISTEN
	  value: "0.0.0.0:%!d(float64=8080)"
```

If `ingressController.admissionWebhook.port` does not have quotes and:

```
	- name: CONTROLLER_ADMISSION_WEBHOOK_LISTEN
	  value: "0.0.0.0:%!d(string=8080)"
```

If `ingressController.admissionWebhook.port` has quotes.

In both cases it has been rendered incorrectly.

To fix, this commit forces the value to string and that is then used to construct
the environment variable.